# src/main/resources/static/openapi-contract.yaml
openapi: 3.0.3

info:
  title: poc-srv-collectors-titleregistration
  description: PoC project for Collector's Channel Title Registration API
  version: 1.0.0

servers:
  - url: http://localhost:8081

tags:
  - name: Registration
    description: Registration operations

paths:
  "/registration/v1/titles":
    get:
      operationId: titleFindAll
      summary: Lists all registered titles with pagination
      tags:
        - Registration
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Titles listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitlePageWrapper'
        'default':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: titleCreate
      summary: Registers a title
      tags:
        - Registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TitleCreateWrapper'
      responses:
        '201':
          description: Title created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleDataWrapper'
        'default':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  "/registration/v1/titles/{titleId}":
    get:
      operationId: titleFindById
      summary: Finds a title by ID
      tags:
        - Registration
      parameters:
        - name: titleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Title found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleDataWrapper'
        '404':
          description: Title not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        'default':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: titleUpdate
      summary: Updates a title by ID
      tags:
        - Registration
      parameters:
        - name: titleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TitleUpdateWrapper'
      responses:
        '200':
          description: Title updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleDataWrapper'
        '404':
          description: Title not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        'default':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: titleDelete
      summary: Deletes a title by ID
      tags:
        - Registration
      parameters:
        - name: titleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Title deleted successfully
        '404':
          description: Title not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        'default':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    TitleCreateWrapper:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/TitleCreateRequest'

    TitleUpdateWrapper:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/TitleUpdateRequest'

    TitleDataWrapper:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TitleCreatedResponse'
        _links:
          $ref: '#/components/schemas/Links'

    TitlePageWrapper:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TitleCreatedResponse'
        pagination:
          $ref: '#/components/schemas/PaginationResponse'
        _links:
          $ref: '#/components/schemas/Links'

    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'

    Link:
      type: object
      properties:
        href:
          type: string
          example: 'http://localhost:8081/registration/v1/titles/0dc42d37-eb5c-4287-bc51-128260d0ef25'

    PaginationResponse:
      type: object
      properties:
        page:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 50
        totalPages:
          type: integer
          example: 5

    TitleCreateRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          example: 'Princess Mononoke'
        originalName:
          type: string
          example: 'Mononoke-hime'
        studio:
          type: string
          example: 'Studio Ghibli'
        type:
          $ref: '#/components/schemas/TitleType'
        barcode:
          type: string
          example: '825553181548'
        purchaseDetails:
          $ref: '#/components/schemas/PurchaseDetailsRequest'

    TitleUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: 'Princess Mononoke'
        originalName:
          type: string
          example: 'Mononoke-hime'
        studio:
          type: string
          example: 'Studio Ghibli'
        type:
          $ref: '#/components/schemas/TitleType'
        barcode:
          type: string
          example: '825553181548'
        purchaseDetails:
          $ref: '#/components/schemas/PurchaseDetailsRequest'

    PurchaseDetailsRequest:
      type: object
      required:
        - store
        - price
        - currency
      properties:
        store:
          type: string
          example: 'Amazon'
        price:
          type: number
          format: float
          example: 125.39
        currency:
          $ref: '#/components/schemas/MonetaryType'

    TitleCreatedResponse:
      type: object
      properties:
        titleId:
          type: string
          format: uuid
          example: '0dc42d37-eb5c-4287-bc51-128260d0ef25'
        createdDateTime:
          type: string
          format: date-time
          example: "2026-08-23T10:00:00Z"
        name:
          type: string
          example: 'Princess Mononoke'
        originalName:
          type: string
          example: 'Mononoke-hime'
        studio:
          type: string
          example: 'Studio Ghibli'
        type:
          $ref: '#/components/schemas/TitleType'
        barcode:
          type: string
          example: '825553181548'
        purchaseDetails:
          $ref: '#/components/schemas/PurchaseDetailsResponse'

    PurchaseDetailsResponse:
      type: object
      required:
        - store
        - price
        - currency
      properties:
        store:
          type: string
          example: 'Amazon'
        price:
          type: number
          format: float
          example: 125.39
        currency:
          $ref: '#/components/schemas/MonetaryType'

    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'

    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        title:
          type: string
        detail:
          type: string

    TitleType:
      type: string
      enum:
        - VHS
        - DVD
        - BLU_RAY
        - BLU_RAY_4K
        - COMBO
      example: BLU_RAY

    MonetaryType:
      type: string
      enum:
        - BRAZILIAN_REAL
        - DOLLAR
        - IENE
      example: BRAZILIAN_REAL